Definitions work
