+++ client/client.ios.ts
@@ -27,7 +27,35 @@ export class MobileServiceClient extends common.MobileServiceClient {
     public getTable(tableName: string): MobileServiceTable {
         return new MobileServiceTable(this._msClient.tableWithName(tableName));
     }
-    public login(provider: string) {
-        console.log("login");
+    public login(provider: string): Promise<MobileServiceUser> {
+        return new Promise((resolve, reject) => {
+            try {
+                /* this._msClient.login((queryResult, error) => {
+                    if (error) {
+                        reject(new Error(error.localizedDescription));
+                        return;
+                    }
+
+                    resolve(utils.getJsObject(queryResult.items));
+                });*/
+                console.log("ios trying");
+            }
+            catch (e) {
+                reject(e);
+            }
+        });
     }
 }
+
+/*        
+        
+         return new Promise((resolve, reject) => {
+            try {
+                let futureResult = this._msClient; // .login(com.microsoft.windowsazure.mobileservices.authentication.MobileServiceAuthenticationProvider[provider]);
+                // utils.futureToPromise(futureResult).then((result) => { resolve(utils.getJsObject(result)); }, reject);
+            }
+            catch (e) {
+                reject(e);
+            }
+        });
+        */
